feat: Add option to exlude first result from a range query
For some queries, the trace server accept a startTime parameter.
On the first run, this parameter is the date of the first trip and on
the next runs, it corresponds to the date of the last saved trip.
Depending on the situation, we want to exclude the first return result
as it will include the trip on this exact date.