feat: Save data by chunks and handle timeout
The connector used to fetch all trips it could find and save them at the
end.
However, this can trigger a timeout when there are a lot of data to
process. Therefore, we now split the trips in chunks and save the whole
chunk in bulk, which is much faster.
Furthermore, we check at the end of each chunk if there are still time
to process the next one before having a timeout. If no, we stop the
execution and start a new job.