fix: Remove duplicates from manual entries
When fetching for manual entries, duplicates can be found when the user
edit the same trip several times.
This led to conflicts, as we were trying to save the same trip more than
once.
Hence, we detect such duplicates and only keep the more recent one.
We also now save sequentially rather than in bulk, as performance is not
critical for manual entries and it makes error messages more explicit.